\name{BBI_WigRegion}
\alias{BBI_WigRegion}
\docType{package}
\title{
Obtains Wig data from a big wig file
}
\description{
Obtains Wig data from a big wig file
}
\usage{
BBI_WigRegion(fileName, chrName, chrStart, chrEnd)	
}
\examples{
\dontrun{
BBI_WigRegion(fileName, chrName, chrStart, chrEnd)	
}
}
